fixdelete feature & import barang

+++ application/controllers/Barang.php
@@ -48,9 +48,9 @@ class Barang extends CI_Controller
         $row->hargabeli,
         $row->hargajual,
         $row->stok,
-        '<a href="'.base_url('barang/view/' . $row->id_barang).'" class="btn btn-primary"><i class="fa fa-eye"></i></a>
-        <a href="'.base_url('barang/edit/' . $row->id_barang).'" class="btn btn-warning"><i class="fa fa-edit"></i></a>
-        <button class="delete-button btn btn-danger" row-data="barang-'.$row->id_barang.'" data-url="'.base_url('barang/delete/' . $row->id_barang).'">
+        '<a href="' . base_url('barang/view/' . $row->id_barang) . '" class="btn btn-primary"><i class="fa fa-eye"></i></a>
+        <a href="' . base_url('barang/edit/' . $row->id_barang) . '" class="btn btn-warning"><i class="fa fa-edit"></i></a>
+        <button class="delete-button btn btn-danger" row-data="barang-' . $row->id_barang . '" data-url="' . base_url('barang/delete/' . $row->id_barang) . '">
             <i class="fa fa-trash"></i>
         </button>'
       );
@@ -167,42 +167,51 @@ class Barang extends CI_Controller
   public function delete($id)
   {
     $barang = $this->barang->getById($id);
-    unlink(FCPATH . 'assets/uploads/barang/' . $barang->gbr);
     $this->barang->delete($id);
+    if($barang->gbr != '' && file_exists('assets/uploads/barang/'.$barang->gbr)){
+      unlink(FCPATH . 'assets/uploads/barang/' . $barang->gbr);
+    }
+    header('Content-Type: application/json');
+    echo json_encode(['section' => 'Barang', 'data' => $barang->nama_barang]);
   }
 
   public function import()
   {
-    $file_mimes = array('application/octet-stream', 'application/vnd.ms-excel', 'application/x-csv', 'text/x-csv', 'text/csv', 'application/csv', 'application/excel', 'application/vnd.msexcel', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    if(isset($_FILES['excelbarang']['name'])) {
+    $allowFile = ['xls', 'xlsx', 'csv', 'xlt'];
+    if (isset($_FILES['excelbarang']['name'])) {
       $arr_file = explode('.', $_FILES['excelbarang']['name']);
       $extension = end($arr_file);
-    
-      if('csv' == $extension) {
+      if (in_array($extension, $allowFile)) {
+        if ('csv' == $extension) {
           $reader = new \PhpOffice\PhpSpreadsheet\Reader\Csv();
           $kodebarang = 0;
           $barang = 1;
-      } else {
+        } else {
           $reader = new \PhpOffice\PhpSpreadsheet\Reader\Xlsx();
           $kodebarang = 1;
           $barang = 2;
-      }
-    
-      $spreadsheet = $reader->load($_FILES['excelbarang']['tmp_name']);
-      $sheetData = $spreadsheet->getActiveSheet()->toArray();
+        }
 
-      for($i = 0;$i < count($sheetData);$i++){
-        $data = [
-          'kodebarang' => $sheetData[$i][$kodebarang],
-          'nama_barang' => $sheetData[$i][$barang],
-          'user' => '0',
-          'id_kategori' => 1
-        ];
-        $this->barang->save($data);
+        $spreadsheet = $reader->load($_FILES['excelbarang']['tmp_name']);
+        $sheetData = $spreadsheet->getActiveSheet()->toArray();
+
+        for ($i = 0; $i < count($sheetData); $i++) {
+          $data = [
+            'kodebarang' => $sheetData[$i][$kodebarang],
+            'nama_barang' => $sheetData[$i][$barang],
+            'user' => '0',
+            'id_kategori' => 1
+          ];
+          $this->barang->save($data);
+        }
+        $this->session->set_flashdata('success', 'Data barang berhasil diimport');
+        redirect(base_url('barang'));
+      } else {
+        $this->session->set_flashdata('error', 'Tipe file tidak sesuai, gunakan file dengan format xls, xlsx, xlt atau csv');
+        redirect(base_url('barang'));
       }
-      redirect(base_url('barang'));
     }
-  } 
+  }
 }
